+++ R/generics.R
@@ -766,9 +766,14 @@ as_tensor.default <- function(x, dtype = NULL, ..., shape = NULL, name = NULL) {
     }
   }
 
+  # doubles get cast to float32, which is inconsistent with the dtype
+  # of other double vector conversions. For consistency, cast to float64
+  if(is.double(x) && !is.array(x))
+    x <- tf$convert_to_tensor(x, name=name, dtype = "float64")
+  else
   # dtype_hint() arg in convert_to_tensor() calls tf$constant(),
   # can silently overflow e.g., tf$convert_to_tensor(-1L, "uint8") --> 255
-  x <- tf$convert_to_tensor(x, name = name)
+    x <- tf$convert_to_tensor(x, name = name)
 
   if (!is.null(dtype)) {
     dtype <- tf$as_dtype(dtype)